Make the script work correctly when there's Reels tab instead of the Search tab
