+++ src/counters_parser.py
@@ -1,4 +1,4 @@
-from src.language_switcher import switch_to_english, LanguageChangedException
+from src.navigation import switch_to_english, LanguageChangedException
 from src.utils import *
